requirements.txt
discord.py
python-dotenv
Flask==2.3.3
requests